Allow deletion

+++ tools/views.py
@@ -1,6 +1,5 @@
 from django.contrib import messages
 from django.core.exceptions import ImproperlyConfigured
-from django.core.urlresolvers import reverse
 from django.http import HttpResponseRedirect
 from django.utils.functional import cached_property
 from django.utils.translation import ugettext as _
@@ -80,17 +79,36 @@ class UpdateView(ToolsMixin, vanilla.UpdateView):
 
 
 class DeleteView(ToolsMixin, vanilla.DeleteView):
+    def allow_delete(self, silent=False):
+        if not silent:
+            messages.error(
+                self.request,
+                _('Deletion of %(class)s "%(object)s" is not allowed.') % {
+                    'class': self.object._meta.verbose_name,
+                    'object': self.object,
+                })
+
+        return False
+
+    def get(self, request, *args, **kwargs):
+        self.object = self.get_object()
+        if not self.allow_delete(silent=False):
+            return HttpResponseRedirect(self.object.get_absolute_url())
+        context = self.get_context_data()
+        return self.render_to_response(context)
+
     def post(self, request, *args, **kwargs):
         self.object = self.get_object()
+        if not self.allow_delete(silent=True):
+            return HttpResponseRedirect(self.object.get_absolute_url())
         self.object.delete()
+        return self.success()
+
+    def success(self):
         messages.success(
             self.request,
             _('%(class)s "%(object)s" has been successfully deleted.') % {
-                'class': self.object._meta.verbose_name,
+                'class': self.model._meta.verbose_name,
                 'object': self.object,
             })
-        return HttpResponseRedirect(self.get_success_url())
-
-    def get_success_url(self):
-        opts = self.model._meta
-        return reverse('%s_%s_list' % (opts.app_label, opts.model_name))
+        return HttpResponseRedirect(self.model().urls.url('list'))
